Fix IndexOutOfRangeException in Series.
The initial implementation of the Series type produced
IndexOutOfRangeExceptions by failing to pass all the arugments from the
Func<T, T,..> constructors to the Func<T[], T> implementation.

This commit reduces the interface to include up to 5 argument
constructors, scaffolds the implementation more effectively to simplify
the Func<T...> constructors, and adds tests to each constructor.